refactor: split mirrorstack.go into topical files, dedupe MCP projections
mirrorstack.go had grown to 977 lines as a catch-all for DB, resources,
discovery, MCP, and core module code. Split into per-concept files that
mirror the existing event.go/cron.go/task.go convention in this package.

Files extracted (same package, no public API change):

  db.go         200 lines — DB/Tx/ModuleDB/ModuleTx + resolvePool helpers +
                lifecycleTxRunner + moduleSchema + package-level wrappers.
  resources.go  108 lines — Cache/Storage/Meter + resolvers + wrappers.
  describe.go   118 lines — Describe/DependsOn/Needs/Resolve + parseDepSpec
                + wrappers. parseDepSpec now uses strings.Cut instead of
                strings.Index (idiomatic since Go 1.18).
  mcp.go        102 lines — MCPTool/MCPResource generics + deriveSchema +
                wrapMCPToolHandler + wrapMCPResourceHandler.

mirrorstack.go shrinks to 486 lines, retaining Config, Module, New, the
scope entry points (Platform/Public/Internal/RequirePermission),
scopedRoutes, Start/startTaskWorker, Close, mountSystemRoutes, and the
convenience-API init block (defaultModule/mustDefault/Init/DefaultModule).

Also fixes a go-reviewer nit: internalRouteBodyCap const was physically
wedged inside the scopedRoutes godoc block, breaking the prose. Moved
above the scopedRoutes comment as its own declaration with its own godoc.

Dedupe in system/:

  system/mcp.go — extract unexported toolEntries() and resourceEntries()
  helpers. Both list handlers and system/manifest.go's buildManifestMCP
  now share the projection logic (was duplicated field-for-field across
  three call sites).

go test ./... + go vet clean. Public API unchanged; no caller migration
needed.
