+++ system/manifest.go
@@ -63,26 +63,14 @@ type ManifestEvents struct {
 	Subscribes map[string]string `json:"subscribes"`
 }
 
-// buildManifestMCP translates the registry's MCP declarations into the
-// wire-safe entries (Handler stripped). Empty slices are non-nil so the
-// output is always {"tools":[],"resources":[]} rather than nulls.
+// buildManifestMCP projects the registry's MCP declarations into wire-safe
+// entries (Handler stripped). Uses the shared toolEntries/resourceEntries
+// helpers from mcp.go so list endpoints and manifest stay in lockstep.
 func buildManifestMCP(reg *registry.Registry) ManifestMCP {
-	tools := reg.MCPTools()
-	toolOut := make([]MCPToolEntry, len(tools))
-	for i, t := range tools {
-		toolOut[i] = MCPToolEntry{
-			Name: t.Name, Description: t.Description,
-			InputSchema: t.InputSchema, OutputSchema: t.OutputSchema,
-		}
+	return ManifestMCP{
+		Tools:     toolEntries(reg.MCPTools()),
+		Resources: resourceEntries(reg.MCPResources()),
 	}
-	resources := reg.MCPResources()
-	resourceOut := make([]MCPResourceEntry, len(resources))
-	for i, rc := range resources {
-		resourceOut[i] = MCPResourceEntry{
-			Name: rc.Name, Description: rc.Description, Schema: rc.Schema,
-		}
-	}
-	return ManifestMCP{Tools: toolOut, Resources: resourceOut}
 }
 
 // ManifestHandler returns an http.HandlerFunc that serves the module manifest.
